use generic slick conversion for AnyVal types
